@@ -252,18 +252,14 @@
             <version>2.4.3</version>
           </plugin>
         </plugins>
-      </pluginManagement>    
+      </pluginManagement>
 	    <plugins>
 	      <plugin>
 	        <groupId>org.apache.maven.plugins</groupId>
 	        <artifactId>maven-compiler-plugin</artifactId>
 	        <configuration>
-	          <source>1.3</source>
-	          <target>1.3</target>
-            <verbose>false</verbose>
-            <fork>true</fork>
-            <executable>${JDK4_HOME}/bin/javac</executable>
-            <compilerVersion>1.3</compilerVersion>	          
+	          <source>1.8</source>
+	          <target>1.8</target>
 	        </configuration>
 	      </plugin>
         <plugin>
